Rename to nflx instead of netflix_ext
