--- metaflow-netflixext/metaflow_extensions/netflix_ext/toplevel/mfextinit_netflixext.py
@@ -1,1 +1,0 @@
-toplevel = "netflixext_toplevel"
