+++ metaflow-netflixext/metaflow_extensions/nflx/plugins/conda/utils.py
@@ -59,8 +59,8 @@ from requests import PreparedRequest, Response, Session
 from requests.auth import AuthBase, HTTPBasicAuth
 
 if TYPE_CHECKING:
-    import metaflow_extensions.netflix_ext.plugins.conda.env_descr
-    from metaflow_extensions.netflix_ext.plugins.conda.env_descr import (
+    import metaflow_extensions.nflx.plugins.conda.env_descr
+    from metaflow_extensions.nflx.plugins.conda.env_descr import (
         PackageSpecification,
         ResolvedEnvironment,
     )
@@ -570,7 +570,7 @@ def is_alias_mutable(alias_type: AliasType, resolved_alias: str) -> bool:
 
 def dict_to_tstr(
     deps: Dict[str, List[str]],
-) -> List["metaflow_extensions.netflix_ext.plugins.conda.env_descr.TStr"]:
+) -> List["metaflow_extensions.nflx.plugins.conda.env_descr.TStr"]:
     from .env_descr import TStr  # Avoid circular import
 
     result = []  # type: List[TStr]
@@ -580,7 +580,7 @@ def dict_to_tstr(
 
 
 def tstr_to_dict(
-    deps: List["metaflow_extensions.netflix_ext.plugins.conda.env_descr.TStr"],
+    deps: List["metaflow_extensions.nflx.plugins.conda.env_descr.TStr"],
 ) -> Dict[str, List[str]]:
     result = {}  # type: Dict[str, List[str]]
     for dep in deps:
