+++ metaflow-netflixext/metaflow_extensions/nflx/toplevel/mfextinit_netflixext.py
@@ -1,0 +1,1 @@
+toplevel = "nflxext_toplevel"
